chore: organized server files into new, well named, folders
