+++ server/routes/basic-routes.js
@@ -1,6 +1,6 @@
 var express = require('express');
 var router = express.Router();
-const queries = require('../queries');
+const queries = require('../queries/basic-queries');
 
 router.use(express.json());
